doctor_appointment_api/requirements.txt
fastapi
uvicorn
psycopg[binary]
sqlalchemy
sqlalchemy[asyncio]
alembic
passlib[bcrypt]
python-jose
pydantic
pydantic-settings
pytest
httpx
asyncpg
